Added NMS for 1.20
Functional Recipes are no longer supported!
