+++ core/src/main/java/com/wolfyscript/utilities/paper/WolfyUtilsBootstrap.java
@@ -1,14 +1,16 @@
 package com.wolfyscript.utilities.paper;
 
+import io.papermc.paper.plugin.bootstrap.BootstrapContext;
 import io.papermc.paper.plugin.bootstrap.PluginBootstrap;
 import io.papermc.paper.plugin.bootstrap.PluginProviderContext;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.jetbrains.annotations.NotNull;
 
 public class WolfyUtilsBootstrap implements PluginBootstrap {
+
     @Override
-    public void bootstrap(@NotNull PluginProviderContext context) {
-        // Do nothing yet
+    public void bootstrap(@NotNull BootstrapContext context) {
+
     }
 
     @Override
