+++ core/src/main/java/me/wolfyscript/utilities/util/reflection/InventoryUpdate.java
@@ -42,7 +42,10 @@ import java.util.Arrays;
 /**
  * A utility class to update the inventory of a player.
  * This is useful to change the title of an inventory.
+ *
+ * @deprecated Spigot introduced the InventoryView#setTitle(String) method in 1.20 rendering this redundant.
  */
+@Deprecated
 public final class InventoryUpdate {
 
     // Classes.
@@ -115,14 +118,26 @@ public final class InventoryUpdate {
     /**
      * Update the player inventory, so you can change the title.
      *
+     * @deprecated Spigot introduced the InventoryView#setTitle(String) method in 1.20 rendering this redundant.
      * @param player   whose inventory will be updated.
      * @param newTitle the new title for the inventory.
      */
+    @Deprecated
     public static void updateInventory(JavaPlugin plugin, Player player, String newTitle) {
         Preconditions.checkNotNull(player, "Cannot update inventory to null player.");
         updateInventory(plugin, player, Component.text(newTitle == null ? "" : newTitle));
     }
 
+    /**
+     * Update the player inventory, so you can change the title.
+     *
+     * @deprecated Spigot introduced the InventoryView#setTitle(String) method in 1.20 rendering this redundant.
+     * @param plugin   The plugin that caused the update.
+     * @param player   whose inventory will be updated.
+     * @param newTitle the new title for the inventory.
+
+     */
+    @Deprecated
     public static void updateInventory(JavaPlugin plugin, Player player, Component newTitle) {
         Preconditions.checkNotNull(player, "Cannot update inventory to null player.");
         try {
