+++ nmsutil/nmsutil-artifact/pom.xml
@@ -32,16 +32,6 @@
     <dependencies>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>nmsutil-v1_16_R1</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>nmsutil-v1_16_R2</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
             <artifactId>nmsutil-v1_16_R3</artifactId>
             <version>${project.version}</version>
         </dependency>
@@ -83,6 +73,11 @@
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>nmsutil-v1_19_R3</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>nmsutil-v1_20_R1</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
