+++ nmsutil/nmsutil-v1_20_R1/src/main/java/me/wolfyscript/utilities/api/nms/v1_20_R1/NBTTagImpl.java
@@ -1,0 +1,115 @@
+/*
+ *       WolfyUtilities, APIs and Utilities for Minecraft Spigot plugins
+ *                      Copyright (C) 2021  WolfyScript
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package me.wolfyscript.utilities.api.nms.v1_20_R1;
+
+import me.wolfyscript.utilities.api.nms.NBTTag;
+import me.wolfyscript.utilities.api.nms.nbt.NBTCompound;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagByte;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagByteArray;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagDouble;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagEnd;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagFloat;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagInt;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagIntArray;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagList;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagLong;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagLongArray;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagShort;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagString;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagByteArrayImpl;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagByteImpl;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagCompoundImpl;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagDoubleImpl;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagEndImpl;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagFloatImpl;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagIntArrayImpl;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagIntImpl;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagListImpl;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagLongArrayImpl;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagLongImpl;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagShortImpl;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.nbt.NBTTagStringImpl;
+
+public class NBTTagImpl extends NBTTag {
+
+    @Override
+    public NBTTagEnd end() {
+        return NBTTagEndImpl.of();
+    }
+
+    @Override
+    public NBTCompound compound() {
+        return NBTTagCompoundImpl.of();
+    }
+
+    @Override
+    public NBTTagList list() {
+        return NBTTagListImpl.of();
+    }
+
+    @Override
+    public NBTTagByte ofByte(byte value) {
+        return NBTTagByteImpl.of(value);
+    }
+
+    @Override
+    public NBTTagByteArray ofByteArray(byte[] value) {
+        return NBTTagByteArrayImpl.of(value);
+    }
+
+    @Override
+    public NBTTagDouble ofDouble(double value) {
+        return NBTTagDoubleImpl.of(value);
+    }
+
+    @Override
+    public NBTTagFloat ofFloat(float value) {
+        return NBTTagFloatImpl.of(value);
+    }
+
+    @Override
+    public NBTTagInt ofInt(int value) {
+        return NBTTagIntImpl.of(value);
+    }
+
+    @Override
+    public NBTTagIntArray ofIntArray(int[] array) {
+        return NBTTagIntArrayImpl.of(array);
+    }
+
+    @Override
+    public NBTTagLong ofLong(long value) {
+        return NBTTagLongImpl.of(value);
+    }
+
+    @Override
+    public NBTTagLongArray ofLongArray(long[] array) {
+        return NBTTagLongArrayImpl.of(array);
+    }
+
+    @Override
+    public NBTTagShort ofShort(short value) {
+        return NBTTagShortImpl.of(value);
+    }
+
+    @Override
+    public NBTTagString ofString(String value) {
+        return NBTTagStringImpl.of(value);
+    }
+}
