+++ nmsutil/nmsutil-v1_20_R1/src/main/java/me/wolfyscript/utilities/api/nms/v1_20_R1/NetworkUtilImpl.java
@@ -1,0 +1,42 @@
+/*
+ *       WolfyUtilities, APIs and Utilities for Minecraft Spigot plugins
+ *                      Copyright (C) 2021  WolfyScript
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package me.wolfyscript.utilities.api.nms.v1_20_R1;
+
+import io.netty.buffer.ByteBuf;
+import me.wolfyscript.utilities.api.nms.NMSUtil;
+import me.wolfyscript.utilities.api.nms.NetworkUtil;
+import me.wolfyscript.utilities.api.nms.network.MCByteBuf;
+import me.wolfyscript.utilities.api.nms.v1_20_R1.network.MCByteBufImpl;
+
+public class NetworkUtilImpl extends NetworkUtil {
+
+    protected NetworkUtilImpl(NMSUtil nmsUtil) {
+        super(nmsUtil);
+    }
+
+    @Override
+    public MCByteBuf buffer(ByteBuf byteBuf) {
+        return new MCByteBufImpl(byteBuf);
+    }
+
+    @Override
+    public MCByteBuf buffer() {
+        return new MCByteBufImpl();
+    }
+}
