+++ nmsutil/nmsutil-v1_20_R1/src/main/java/me/wolfyscript/utilities/api/nms/v1_20_R1/nbt/NBTTagTypeImpl.java
@@ -1,0 +1,49 @@
+/*
+ *       WolfyUtilities, APIs and Utilities for Minecraft Spigot plugins
+ *                      Copyright (C) 2021  WolfyScript
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package me.wolfyscript.utilities.api.nms.v1_20_R1.nbt;
+
+import java.util.function.Function;
+import me.wolfyscript.utilities.api.nms.nbt.NBTBase;
+import me.wolfyscript.utilities.api.nms.nbt.NBTTagType;
+import net.minecraft.nbt.Tag;
+import org.jetbrains.annotations.Nullable;
+
+public class NBTTagTypeImpl<T extends NBTBase> implements NBTTagType<T> {
+
+    protected final Type type;
+    private final Function<Object, T> function;
+
+    public NBTTagTypeImpl(Type type, Function<Object, T> nbtBase) {
+        this.type = type;
+        this.function = nbtBase;
+    }
+
+    public T get(@Nullable Object nbtBase) {
+        if (nbtBase instanceof Tag && type.is(((Tag) nbtBase).getId())) {
+            return function.apply(nbtBase);
+        }
+        return null;
+    }
+
+    @Override
+    public Type getType() {
+        return type;
+    }
+
+}
