+++ pom.xml
@@ -34,7 +34,7 @@
         <plugin_version.specialsource>1.2.4</plugin_version.specialsource>
         <plugin_version.license>2.0.0</plugin_version.license>
         <!-- The dependency versions -->
-        <depend_version.spigotapi>1.19.3-R0.1-SNAPSHOT</depend_version.spigotapi>
+        <depend_version.spigotapi>1.20-R0.1-SNAPSHOT</depend_version.spigotapi>
         <depend_version.mojang-authlib>1.5.21</depend_version.mojang-authlib>
         <depend_version.jackson>2.14.0-rc1</depend_version.jackson>
         <depend_version.jetbrains-annotations>23.0.0</depend_version.jetbrains-annotations>
@@ -45,7 +45,7 @@
         <depend_version.adventure-api>4.11.0</depend_version.adventure-api>
         <depend_version.adventure-platform>4.1.2</depend_version.adventure-platform>
         <depend_version.adventure-minimessage>4.11.0</depend_version.adventure-minimessage>
-        <depend_version.nbt-api>2.11.2</depend_version.nbt-api>
+        <depend_version.nbt-api>2.11.3</depend_version.nbt-api>
         <depend_version.javassist>3.29.1-GA</depend_version.javassist>
     </properties>
